Tick 1 -> 0.5 증가로 변경
더 자세한 애니메이션 + 1틱 문제 해결

+++ Assets/Scripts/Animation/UI/AnimPanel.cs
@@ -69,9 +69,9 @@ namespace Animation.UI
                 tickSpeedField.text = _manager.TickSpeed.ToString();
         }
 
-        private void AnimManager_TickChanged(int obj)
+        private void AnimManager_TickChanged(float obj)
         {
-            tickField.text = obj.ToString();
+            tickField.text = ((int)obj).ToString();
         }
 
         public void Stop()
